op-service,batcher,proposer: Retry when dialing RPC URLs
This adds retries when dialing ethclient or rollupclient RPC urls.
Because this now does up to 10 exponential backoffs, the dial timeout
has been increased to 15s (though this may not be enough).

I have also modified the API to DialClient to no longer take a context.
This is because we are not sharing the context & if we want to pass in
a context, we should probably set the timeout outside the dial function.